rename theme file, add Kotlin, fix one ObjC block
